test: promote Utils and OpenTelemetryProvider mixins into standalone unit test classes
